+++ NEWS.md
@@ -1,6 +1,10 @@
 # ompr 0.6.0.9000
 
-* ...
+...
+
+## Breaking changes
+
+* ompr now uses sparse constraint matrices. `extract_constraints` now returns a sparse matrix and `objective_function` returns a sparse vector.
 
 ## Minor changes